Move Compiler to compilation directory and split items
